requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.32.0
django-rest-auth==0.9.1
djangorestframework==3.6.3
httpie==0.9.9
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
oauthlib==2.0.2
packaging==16.8
Pygments==2.2.0
PyMySQL==0.7.11
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
